Removing events_manager module
