+++ src/app/beer_garden/db/api.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import beer_garden.db.mongo.api
-from beer_garden.events.events_manager import publish_event
+from beer_garden.events import publish_event
 from brewtils.models import Events
 
 check_connection = beer_garden.db.mongo.api.check_connection
