+++ src/app/beer_garden/plugin.py
@@ -11,7 +11,7 @@ import beer_garden
 import beer_garden.config
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
-from beer_garden.events.events_manager import publish_event
+from beer_garden.events import publish_event
 
 logger = logging.getLogger(__name__)
