+++ src/app/beer_garden/queues.py
@@ -5,7 +5,7 @@ from brewtils.models import Events, Queue, System
 
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
-from beer_garden.events.events_manager import publish_event
+from beer_garden.events import publish_event
 from beer_garden.queue.rabbit import get_routing_key
 
 logger = logging.getLogger(__name__)
