+++ src/app/beer_garden/requests.py
@@ -24,8 +24,7 @@ from requests import Session
 import beer_garden.config
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
-from beer_garden.errors import RoutingRequestException
-from beer_garden.events.events_manager import publish_event
+from beer_garden.events import publish_event
 from beer_garden.metrics import request_completed, request_created, request_started
 
 logger = logging.getLogger(__name__)
